# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Projects Tracker API
  description: |
    RESTful API for the Projects Tracker application. This API provides endpoints for managing
    projects, vendors, budgets, divisions, and related entities.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the login endpoint. Add it to the Authorization header
        as "Bearer <token>".
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
      required:
        - message
        - code
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
paths:
...
  /settings/public/{key}:
    get:
      tags:
        - Settings
      summary: Get a public setting
      description: Get a single application setting by key. Available to all authenticated users.
      parameters:
        - name: key
          in: path
          required: true
          description: Setting key (e.g. avatar_style)
          schema:
            type: string
      responses:
        '200':
          description: Setting value retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      key:
                        type: string
                      value:
                        type: string
                        nullable: true
        '401':
          description: Not authenticated
  /settings/{key}:
    put:
      tags:
        - Settings
      summary: Update setting
      description: Update an application setting by key (superadmin only)
      parameters:
        - name: key
          in: path
          required: true
          description: Setting key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - value
              properties:
                value:
                  type: string
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          description: Setting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== Audit Logs ====================
...